Use marshaler for json logging.
