+++ exporters/logging-otlp/src/main/java/io/opentelemetry/exporter/logging/otlp/OtlpJsonLoggingMetricExporter.java
@@ -5,33 +5,24 @@
 
 package io.opentelemetry.exporter.logging.otlp;
 
-import static io.opentelemetry.exporter.logging.otlp.HexEncodingStringJsonGenerator.JSON_FACTORY;
+import static io.opentelemetry.exporter.logging.otlp.JsonUtil.JSON_FACTORY;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.core.io.SegmentedStringWriter;
-import io.opentelemetry.exporter.otlp.internal.MetricAdapter;
-import io.opentelemetry.proto.metrics.v1.ResourceMetrics;
+import io.opentelemetry.exporter.otlp.internal.metrics.ResourceMetricsMarshaler;
 import io.opentelemetry.sdk.common.CompletableResultCode;
 import io.opentelemetry.sdk.metrics.data.MetricData;
 import io.opentelemetry.sdk.metrics.export.MetricExporter;
 import java.io.IOException;
 import java.util.Collection;
-import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import org.curioswitch.common.protobuf.json.MessageMarshaller;
 
 /**
  * A {@link MetricExporter} which writes {@linkplain MetricData spans} to a {@link Logger} in OTLP
- * JSON format. Each log line will include a single {@link ResourceMetrics}.
+ * JSON format. Each log line will include a single {@code ResourceMetrics}.
  */
 public final class OtlpJsonLoggingMetricExporter implements MetricExporter {
-
-  private static final MessageMarshaller marshaller =
-      MessageMarshaller.builder()
-          .register(ResourceMetrics.class)
-          .omittingInsignificantWhitespace(true)
-          .build();
 
   private static final Logger logger =
       Logger.getLogger(OtlpJsonLoggingMetricExporter.class.getName());
@@ -45,11 +36,11 @@ public final class OtlpJsonLoggingMetricExporter implements MetricExporter {
 
   @Override
   public CompletableResultCode export(Collection<MetricData> metrics) {
-    List<ResourceMetrics> allResourceMetrics = MetricAdapter.toProtoResourceMetrics(metrics);
-    for (ResourceMetrics resourceMetrics : allResourceMetrics) {
+    ResourceMetricsMarshaler[] allResourceMetrics = ResourceMetricsMarshaler.create(metrics);
+    for (ResourceMetricsMarshaler resourceMetrics : allResourceMetrics) {
       SegmentedStringWriter sw = new SegmentedStringWriter(JSON_FACTORY._getBufferRecycler());
-      try (JsonGenerator gen = HexEncodingStringJsonGenerator.create(sw)) {
-        marshaller.writeValue(resourceMetrics, gen);
+      try (JsonGenerator gen = JsonUtil.create(sw)) {
+        resourceMetrics.writeJsonTo(gen);
       } catch (IOException e) {
         // Shouldn't happen in practice, just skip it.
         continue;
