+++ exporters/logging-otlp/src/test/java/io/opentelemetry/exporter/logging/otlp/OtlpJsonLoggingMetricExporterTest.java
@@ -101,7 +101,7 @@ class OtlpJsonLoggingMetricExporterTest {
             + "          \"timeUnixNano\": \"2\","
             + "          \"asDouble\": 4.0"
             + "        }],"
-            + "        \"aggregationTemporality\": \"AGGREGATION_TEMPORALITY_CUMULATIVE\","
+            + "        \"aggregationTemporality\": 2,"
             + "        \"isMonotonic\": true"
             + "      }"
             + "    }]"
@@ -124,7 +124,7 @@ class OtlpJsonLoggingMetricExporterTest {
             + "          \"timeUnixNano\": \"2\","
             + "          \"asDouble\": 4.0"
             + "        }],"
-            + "        \"aggregationTemporality\": \"AGGREGATION_TEMPORALITY_CUMULATIVE\","
+            + "        \"aggregationTemporality\": 2,"
             + "        \"isMonotonic\": true"
             + "      }"
             + "    }]"
