+++ exporters/logging-otlp/src/test/java/io/opentelemetry/exporter/logging/otlp/OtlpJsonLoggingSpanExporterTest.java
@@ -118,12 +118,12 @@ class OtlpJsonLoggingSpanExporterTest {
             + "      \"traceId\": \"12340000000043211234000000004321\","
             + "      \"spanId\": \"8765000000005678\","
             + "      \"name\": \"testSpan2\","
-            + "      \"kind\": \"SPAN_KIND_CLIENT\","
+            + "      \"kind\": 3,"
             + "      \"startTimeUnixNano\": \"500\","
             + "      \"endTimeUnixNano\": \"1501\","
             + "      \"status\": {"
-            + "        \"deprecatedCode\": \"DEPRECATED_STATUS_CODE_UNKNOWN_ERROR\","
-            + "        \"code\": \"STATUS_CODE_ERROR\""
+            + "        \"deprecatedCode\": 2,"
+            + "        \"code\": 2"
             + "      }"
             + "    }]"
             + "  }, {"
@@ -135,7 +135,7 @@ class OtlpJsonLoggingSpanExporterTest {
             + "      \"traceId\": \"12345678876543211234567887654321\","
             + "      \"spanId\": \"8765432112345678\","
             + "      \"name\": \"testSpan1\","
-            + "      \"kind\": \"SPAN_KIND_INTERNAL\","
+            + "      \"kind\": 1,"
             + "      \"startTimeUnixNano\": \"100\","
             + "      \"endTimeUnixNano\": \"1100\","
             + "      \"attributes\": [{"
@@ -160,7 +160,7 @@ class OtlpJsonLoggingSpanExporterTest {
             + "        }]"
             + "      }],"
             + "      \"status\": {"
-            + "        \"code\": \"STATUS_CODE_OK\""
+            + "        \"code\": 1"
             + "      }"
             + "    }]"
             + "  }]"
